Reorganize client files for runtime
* Group the methods in the files so that the public methods come before
  private. I find this makes it much easier to read the files and find
  the relevant parts of the logic.
* Move the error logic for the Docker client into a separate file for
  readability.
